+++ admin/doc-python-common-requirements.txt
@@ -1,3 +1,1 @@
-pcpp
-Jinja2
 src/python-common
