+++ doc/.gitignore
@@ -1,3 +1,2 @@
 /overview.png
 /object_store.png
-/api/mon_command_api.rst
