+++ doc/api/mon_command_api.rst
@@ -1,0 +1,2 @@
+.. ceph-mgr-commands:: src/pybind/mgr
+.. ceph-mon-commands:: src/mon/MonCommands.h src/mgr/MgrCommands.h
